Fixed issues reported

+++ src/helper.js
@@ -22,7 +22,7 @@ exports.checkSource = (path) => {
 
 exports.createFolder = (path) => {
     try {
-        // first check if directory already exists
+        // if folder already exists
         if (!fs.existsSync(path)) {
             fs.mkdirSync(path);
         }
@@ -32,6 +32,9 @@ exports.createFolder = (path) => {
 }
 
 exports.deleteFolder = (path) => {
+    // if folder is already nonexistent 
+    if (this.checkSource(path) == "none")
+        return;
     fs.rmSync(path, { recursive: true}, (err) => {
         if (err) {
             console.log(chalk.bgRed("Error:") + chalk.red(err));
@@ -69,7 +72,11 @@ exports.pasteIntoTemplate = (title, body, stylesheet) => {
     }
     html += `<meta name="viewport" content="width=device-width, initial-scale=1">
         </head>
-        <body style="background-color:#F2EECB;">`;
+        <body`
+    if (stylesheet == '') {
+        html += ` style="background-color:#F2EECB;"`;
+    }
+    html += `>`;
     if (title != '') {
         html += `
             <h1>${title}</h1>`;
